requirements.txt
Django==1.10.2
django-cors-headers==1.2.2
djangorestframework==3.4.7
django-extensions==1.7.4
